requirements.txt
streamlit
numpy
pandas
pillow
torch
faiss-cpu
openai-clip
